+++ tests/integration/test_plugins_usgs.py
@@ -5,9 +5,12 @@ import pytest
 
 from typing import Iterator
 
+from pydantic import ValidationError
+
 from basin3d.core.models import Base
-from basin3d.core.types import FeatureTypes, TimeFrequency, ResultQuality, SamplingMedium
-from basin3d.synthesis import register, SynthesisException, get_timeseries_data
+from basin3d.core.schema.enum import ResultQualityEnum, TimeFrequencyEnum
+from basin3d.core.types import SamplingMedium
+from basin3d.synthesis import register, get_timeseries_data
 
 
 @pytest.mark.integration
@@ -24,7 +27,8 @@ def test_measurement_timeseries_tvp_observations_usgs():
         "aggregation_duration": "DAY",
         "results_quality": "CHECKED"
     }
-    with pytest.raises(SynthesisException):
+
+    with pytest.raises(ValidationError):
         synthesizer.measurement_timeseries_tvp_observations(**query0)
 
     query1 = {
@@ -53,7 +57,8 @@ def test_measurement_timeseries_tvp_observations_usgs():
         "aggregation_duration": "DAY",
         "results_quality": "CHECKED"
     }
-    with pytest.raises(TypeError):
+
+    with pytest.raises(ValidationError):
         synthesizer.measurement_timeseries_tvp_observations(**query2)
 
 
@@ -61,9 +66,10 @@ def test_measurement_timeseries_tvp_observations_usgs():
 @pytest.mark.parametrize("query, feature_type", [({"id": "USGS-13"}, "region"),
                                                  ({"id": "USGS-0102"}, "subregion"),
                                                  ({"id": "USGS-011000"}, "basin"),
-                                                 ({"id": "USGS-01020004"}, "subbasin")], ids=["region", "subregion",
-
-                                                                                              "basin", "subbasin"])
+                                                 ({"id": "USGS-01020004"}, "subbasin"),
+                                                 ({"id": "USGS-09129600"}, "point"),
+                                                 ({"id": "USGS-383103106594200", "feature_type": "POINT"}, "point")],
+                         ids=["region", "subregion", "basin", "subbasin", "point", "point_long_id"])
 def test_usgs_monitoring_feature(query, feature_type):
     """Test USGS search by region  """
 
@@ -72,29 +78,29 @@ def test_usgs_monitoring_feature(query, feature_type):
 
     assert monitoring_feature is not None
     assert isinstance(monitoring_feature, Base)
-    assert FeatureTypes.TYPES[monitoring_feature.feature_type] == feature_type.upper()
+    assert monitoring_feature.id == query["id"]
+    assert monitoring_feature.feature_type == feature_type.upper()
 
 
-# @pytest.mark.integration
-@pytest.mark.parametrize("query, expected_count",
-                         [({"datasource": "USGS"}, 2889),
-                          ({"monitoring_features": ['USGS-02']}, 1),
-                          ({"feature_type": "region"}, 21),
-                          ({"feature_type": "subregion"}, 222),
-                          ({"feature_type": "basin"}, 379),
-                          ({"feature_type": "subbasin"}, 2267),
-                          ({"feature_type": "watershed"}, 0),
-                          ({"feature_type": "subwatershed"}, 0),
-                          ({"feature_type": "site"}, 0),
-                          ({"feature_type": "plot"}, 0),
-                          ({"feature_type": "vertical path"}, 0),
-                          ({"feature_type": "horizontal path"}, 0),
-                          ({"feature_type": "point"}, 0),
-                          ({"monitoring_features": ["USGS-09129600"], "feature_type": "point"}, 1),
-                          ({"parent_features": ['USGS-02']}, 118),
-                          ({"parent_features": ['USGS-02020004'], "feature_type": "point"}, 49),
-                          ({"parent_features": ['USGS-0202'], "feature_type": "subbasin"}, 8),
-                          ({"parent_features": ['USGS-020200'], "feature_type": "point"}, 0)],
+@pytest.mark.integration
+@pytest.mark.parametrize("query, expected_count", [({"datasource": "USGS"}, 2889),
+                                                   ({"monitoring_features": ['USGS-02']}, 1),
+                                                   ({"feature_type": "region"}, 21),
+                                                   ({"feature_type": "subregion"}, 222),
+                                                   ({"feature_type": "basin"}, 379),
+                                                   ({"feature_type": "subbasin"}, 2267),
+                                                   ({"feature_type": "watershed"}, 0),
+                                                   ({"feature_type": "subwatershed"}, 0),
+                                                   ({"feature_type": "site"}, 0),
+                                                   ({"feature_type": "plot"}, 0),
+                                                   ({"feature_type": "vertical path"}, 0),
+                                                   ({"feature_type": "horizontal path"}, 0),
+                                                   ({"feature_type": "point"}, 0),
+                                                   ({"monitoring_features": ["USGS-09129600"], "feature_type": "point"}, 1),
+                                                   ({"parent_features": ['USGS-02']}, 118),
+                                                   ({"parent_features": ['USGS-02020004'], "feature_type": "point"}, 49),
+                                                   ({"parent_features": ['USGS-0202'], "feature_type": "subbasin"}, 8),
+                                                   ({"parent_features": ['USGS-020200'], "feature_type": "point"}, 0)],
                          ids=["all", "region_by_id", "region", "subregion",
                               "basin", "subbasin",
                               "watershed", "subwatershed",
@@ -108,43 +114,16 @@ def test_usgs_monitoring_features(query, expected_count):
     """Test USGS search by region  """
 
     synthesizer = register(['basin3d.plugins.usgs.USGSDataSourcePlugin'])
-    monitoring_featurues = synthesizer.monitoring_features(**query)
+    monitoring_features = synthesizer.monitoring_features(**query)
 
     # TODO should there be some kind of exeption handling for invalid queries that don't return anything?
     count = 0
-    for mf in monitoring_featurues:
+    for mf in monitoring_features:
         count += 1
         print(
-            f"{mf.id} ({FeatureTypes.TYPES[mf.feature_type]}) {mf.description} {mf.coordinates and [(p.x, p.y) for p in mf.coordinates.absolute.horizontal_position]}")
+            f"{mf.id} ({mf.feature_type}) {mf.description} {mf.coordinates and [(p.x, p.y) for p in mf.coordinates.absolute.horizontal_position]}")
         if 'feature_type' in query:
-            assert FeatureTypes.TYPES[mf.feature_type] == query['feature_type'].upper()
-
-    assert count == expected_count
-
-
-@pytest.mark.parametrize("query, expected_count", [
-    ({"feature_type": ['USGS-02']}, 1)],
-                         ids=["region_by_id"])
-def test_usgs_monitoring_features2(query, expected_count, monkeypatch):
-    """Test USGS search by region  """
-
-    synthesizer = register(['basin3d.plugins.usgs.USGSDataSourcePlugin'])
-    monitoring_featurues = synthesizer.monitoring_features(**query)
-
-    # TODO should there be some kind of exeption handling for invalid queries that don't return anything?
-    count = 0
-    print(query.values(), "count:", count, "expected:", expected_count)
-
-    for mf in monitoring_featurues:
-        count += 1
-        print(FeatureTypes.TYPES[mf.feature_type])
-        print(query['feature_type'].upper())
-        print(
-            f"{mf.id} ({FeatureTypes.TYPES[mf.feature_type]}) {mf.description} {mf.coordinates and [(p.x, p.y) for p in mf.coordinates.absolute.horizontal_position]}")
-        if 'feature_type' in query:
-            assert FeatureTypes.TYPES[mf.feature_type] == query['feature_type'].upper()
-
-    print(query.values(), "count:", count, "expected:", expected_count)
+            assert mf.feature_type == query['feature_type'].upper()
 
     assert count == expected_count
 
@@ -152,23 +131,29 @@ def test_usgs_monitoring_features2(query, expected_count, monkeypatch):
 @pytest.mark.integration
 def test_usgs_get_data():
     synthesizer = register(['basin3d.plugins.usgs.USGSDataSourcePlugin'])
-    usgs_df, usgs_metadata = get_timeseries_data(
-        synthesizer=synthesizer, monitoring_features=["USGS-09110000"],
-        observed_property_variables=['RDC', 'WT'], start_date='2019-10-25', end_date='2019-10-28')
+    usgs_data = get_timeseries_data(synthesizer=synthesizer, monitoring_features=["USGS-09110000"],
+                                    observed_property_variables=['RDC', 'WT'], start_date='2019-10-25',
+                                    end_date='2019-10-28')
+    usgs_df = usgs_data.data
+    usgs_metadata_df = usgs_data.metadata
 
     # check the dataframe
     assert isinstance(usgs_df, pd.DataFrame) is True
     for column_name in list(usgs_df.columns):
-        assert column_name in ['TIMESTAMP', 'USGS-09110000__RDC', 'USGS-09110000__WT']
-    assert usgs_df.shape == (4, 3)
-    alpha_data = usgs_df.get('USGS-09110000__RDC')
+        assert column_name in ['TIMESTAMP', 'USGS-09110000__RDC__MEAN', 'USGS-09110000__WT__MEAN',
+                               'USGS-09110000__WT__MIN', 'USGS-09110000__WT__MAX']
+    assert usgs_df.shape == (4, 5)
+    alpha_data = usgs_df.get('USGS-09110000__RDC__MEAN')
     assert list(alpha_data.values) == [4.2475270499999995, 4.219210203, 4.134259662, 4.332477591]
     assert list(alpha_data.index) == [dt.datetime(2019, 10, num) for num in range(25, 29)]
+
     # make sure the temporary data directory is removed
     temp_dir = os.path.join(os.getcwd(), 'temp_data')
     assert os.path.isdir(temp_dir) is False
+
     # check the metadata store
-    var_metadata = usgs_metadata.get('USGS-09110000__RDC')
+    # Get synthesized variable field names and values
+    var_metadata = usgs_metadata_df['USGS-09110000__RDC__MEAN']
     assert var_metadata['data_start'] == dt.datetime(2019, 10, 25)
     assert var_metadata['data_end'] == dt.datetime(2019, 10, 28)
     assert var_metadata['records'] == 4
@@ -176,9 +161,36 @@ def test_usgs_get_data():
     assert var_metadata['basin_3d_variable'] == 'RDC'
     assert var_metadata['basin_3d_variable_full_name'] == 'River Discharge'
     assert var_metadata['statistic'] == 'MEAN'
-    assert var_metadata['temporal_aggregation'] == TimeFrequency.DAY
-    assert var_metadata['quality'] == ResultQuality.RESULT_QUALITY_CHECKED
+    assert var_metadata['temporal_aggregation'] == TimeFrequencyEnum.DAY
+    assert var_metadata['quality'] == ResultQualityEnum.CHECKED
     assert var_metadata['sampling_medium'] == SamplingMedium.WATER
     assert var_metadata['sampling_feature_id'] == 'USGS-09110000'
     assert var_metadata['datasource'] == 'USGS'
     assert var_metadata['datasource_variable'] == '00060'
+
+    assert usgs_metadata_df['USGS-09110000__WT__MIN']['statistic'] == 'MIN'
+    assert usgs_metadata_df['USGS-09110000__WT__MAX']['statistic'] == 'MAX'
+
+    # check filtering by single statistic
+    usgs_data = get_timeseries_data(synthesizer=synthesizer, monitoring_features=["USGS-09110000"],
+                                    observed_property_variables=['RDC', 'WT'], start_date='2019-10-25',
+                                    end_date='2019-10-28', statistic=['MEAN'])
+    usgs_df = usgs_data.data
+
+    # check the dataframe
+    assert isinstance(usgs_df, pd.DataFrame) is True
+    for column_name in list(usgs_df.columns):
+        assert column_name in ['TIMESTAMP', 'USGS-09110000__RDC__MEAN', 'USGS-09110000__WT__MEAN']
+    assert usgs_df.shape == (4, 3)
+
+    # check filtering by multiple statistic
+    usgs_data = get_timeseries_data(synthesizer=synthesizer, monitoring_features=["USGS-09110000"],
+                                    observed_property_variables=['RDC', 'WT'], start_date='2019-10-25',
+                                    end_date='2019-10-28', statistic=['MIN', 'MAX'])
+    usgs_df = usgs_data.data
+
+    # check the dataframe
+    assert isinstance(usgs_df, pd.DataFrame) is True
+    for column_name in list(usgs_df.columns):
+        assert column_name in ['TIMESTAMP', 'USGS-09110000__WT__MIN', 'USGS-09110000__WT__MAX']
+    assert usgs_df.shape == (4, 3)